Add an orderly command line tool.
A lot of the operational APIs in the package were exposed only as
functions, following the R practice. It is more common in Python to
expose these as a command like tool.

This adds basic commands to create a repository, run a report, search
packets and manage locations. More features and bells and whistles can
be added later on.

The tool is currently named `orderly`, following the package's name.
This may change in the future, if and when we decide to rename the
package.